R7.1: split vector/events into model/produce/check
Facade mod.rs re-exports everything; no public path changes. This is
the assurance-boundary module: the intentional producer/checker
duplication is now physically separated — producer derivations
(event_templates, build_order, build_dependencies,
producer_effect_dependencies, producer_checked_sample_count) in
produce.rs; independent checker re-derivations
(independently_expected_event_keys, independently_order_events,
checker_required_effect_dependencies, independent_checked_sample_count)
in check.rs — with module headers forbidding their merge (plan §3.2).

Plan §4.8 — admission guards in the shared verify function
`verify_event_order_certificate_impl` (hosted in check.rs):
- verify_event_table_independently
- verify_required_dependencies
- verify_compact_repetition_basis
- validate_layout / validate_order

remain on BOTH paths after the split: the producer's terminal
verification (build_event_order_certificate,
build_state_event_order_certificate) and the standalone checker entries
(verify_event_order_certificate, verify_state_event_order_certificate).
Existing rejection tests fail through the checker entry alone (mutated
certificates -> VectorOrderMismatch, DependencyMismatch) and stay
green.

add_dependency (total vocabulary helper used by both sides) moved to
model.rs so check never imports from produce (plan §4.6).

Validation: transform 387 lib tests, transform clippy (+1.97.0) clean.
