+++ porting/journal/2026-07-20.md
@@ -105,3 +105,31 @@ modules on both sides of the plan artifact. Validation: transform 385,
   `pub(in crate::signal_fir::vector)` dans la façade (le module `module`
   l'importe par `super::lower::`).
 - Gates : transform 387, clippy 0, fmt vert.
+
+## R7.1 — scission de `vector/events` (frontière d'assurance)
+
+- `events/mod.rs` (2 268 lignes) scindé en `model.rs` (315), `produce.rs`
+  (836), `check.rs` (1 090), façade `mod.rs` (38).
+- La duplication producteur/checker est physiquement séparée et documentée
+  dans les deux en-têtes : côté producteur `event_templates`,
+  `build_order`, `build_dependencies`, `producer_effect_dependencies`,
+  `producer_checked_sample_count` ; côté checker `independently_expected_
+  event_keys`, `independently_order_events`, `checker_required_effect_
+  dependencies`, `independent_checked_sample_count`. Interdiction de
+  fusion notée (plan §3.2).
+- §4.8 : `verify_event_order_certificate_impl` reste dans `check.rs`,
+  appelé par la vérification terminale du producteur (build_*) ET par les
+  entrées checker autonomes. Tests de rejet par l'entrée checker seule
+  déjà présents (mutations ordre/dépendances → VectorOrderMismatch,
+  DependencyMismatch) — inchangés et verts.
+- `add_dependency` (5 lignes, vocabulaire total) déplacé vers `model.rs`
+  pour éviter que check appelle produce (§4.6).
+- Gates : transform 387, clippy 0, fmt vert.
+- Incident (corrigé avant poussée) : `cargo fix` avait mis sous
+  `#[cfg(test)]` trois imports pourtant requis par du code non-test
+  (EffectAtom/ForeignPurity/StateResource dans check.rs et produce.rs,
+  RoutedUseSource/VectorRegion dans model.rs). Le build *test* passait
+  (cfg(test) actif) mais le build lib seul — et donc le compilateur
+  release de l'arbitre — échouait. Détecté par l'échec de l'arbitre,
+  gates re-validées après retrait des gates bidon ; leçon : valider
+  `cargo build -p transform` (lib sans cfg test) après chaque cargo fix.
